+++ CHANGELOG.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.3.0]
+### Added
+- floor id returned on region enter
+- `getPointFromCoordinate` method that takes floor id, lat and lng and translates it to point on the
+map image
+
+### Removed
+- outdated @todo
+
 ## [0.2.0]
 ### Added
 - floor level, floor certainty, altitude and bearing on location change